Added some Tremination Functionality
When the app is terminated and the student is in a conversation, the app will now end the session.

Added if statements for other scenerios such as when connecting to tutor or when connected with tutor but need to fill in these if statements.

Also fixed some minore UI bugs in post a question process.