Implementing termination and the Intermediary screen for student tutor connection
Wrote some code in the appwillterminate of app delegate for student app to complete certain tasks before the app terminates. There is 1 more scenerio that needs to be filled out and, 1 scenerio that is filled but not working. The scenerio that is fillde but not working is if on the tutor connecting screen.

For implenting functions upon termination it is a good idea to look in websockets so that the server will know when the phone disconeects rather than the phone having to send an order stating that its going to discconect upn termination, since operations upon terminations seem risky.

Also implemented intermediary screen when connecting the student and the tutor. Once the student posts a question, and a tutor clicks rescue now the tutor will then go to a screen with an image of the question and a button for when the tutor is ready. The student will go to a screen stating the tutor is working on your problem. Once the tutor clicks Im ready then both the student and tutor will be put into the chat session.