Updated swipe to delete
Swipe to the left to delete and mark as unread is available
Also added support to mark as unread, this is done by mark the last message
as unread and update the session. Then, the table view is redrawn.

For deletion, a alertView action sheet is present to confirm the
deletion. The reason for adding this function is, it's easy to swipe too
much to accidentally delete one row, if there's no confirmation