Fix: photos not showing in the history session
In previous versions, the images in the messages are not copied to the
newly created message. This version distinguishes between the image message
and the text message. For images, create new message with image data.

When the chatVC populates the cells (bubbles) with images, it'll try to download the
image from server if the download status is not successful. Because ideally, the
 thumbnails are stored locally already, so setting the thumbnail download status to
successful for every image. However, the received image could not have been
downloaded yet, so have to distinguish between sent/received image.

Note that, for sent images, the message model will not contain its storage space
in the cloud