fix image loading problem
In previous versions, wrongly used model.image to get image view, but
model might not contain the image view, if the message is incoming.
In this version, if the message is outgoing, just use local image for the
saved image, but for incoming ones, copy its original image body.

Lastly, the messages are not deleted even if its containing conversation is deleted