Update: session saving method
Old way: pull each message out, make a new message and copy the reference of
the received object of that message. And delete a conversation, but not the messages.

Later it was found out that hyphenate locates the messages by conversation ID. if the old
messages from old conversations are not removed, if the new tutor come in with the same
 senderID(which is the conversation ID), the old history will be displayed for the student.
To utilize this point, instead of making new messages for each new message received,
just modify the message's conversation id and store it to a newly created conversation .

In addition, download the images that haven'y been downloaded.

Plus, the sender tag of the cell is removed.