Created new xib and swift for profile page cell
The XIB and swift file is copied from ConversationTableViewCell. Reason: the conversation
tableviewcell is modified on the git tree and this view should not be used in both historyTVC
and the profileVC. The current conversationTVCell will be checked out in this commits to reflect
the change on the dev branch.