Added support for loading the profile at login
The loading will start if the profileNeedsUpdate flag is set. This flag is set when
1. First launch the app
2. A user logs out

When user logs out, it also resets the profile value to defaults once.

The settingTVC is made to be the config.swift's delegate, so that when the profile data
comes back from Firebase, settingTVC is notified to reload data.

Modified the cells:
1. Setting profile table view cell: shrunk the cell height, changed constrint and text size
	removed some useless labels
2. Header view cell, changed size and font size, in MyprofileVCTVC

In shopTVC, fixed that the alertview won't display how many minutes user purchased

For Email/Grade/Name/ProfileSetting VC, changed database strings to constants defined,

Fixed possible crash due to force logout, which looks for main.storyboard(removed way earlier)