Fix crash when loading profile photo
Crash because1. The data stored in user defaults is really Data, not the file name
to the image. 2. the initilization on FireBase set profileimage to an empty string,
and alamo fire just opens it, returning an empty data.

Solution: 1. store the imgaedata. 2. validate the url before connection

Then, fix corner radius issue