Use database value to check if tutor is connected or ready to chat
In last version, we use friend request to trigger notification, then add notification in
summaryVC and tutorConnectedVC to trigger next step. When internet is slow, this method
will cause delay.

Now, when tutor is connected, tutor app will write 1 to status, when tutor is ready to chat, tutor app
will write 2 to status. In student app, add observer to status and based on the value of status, the app
will push different VC.