Bugs fixed
1.
Bug:
In summaryVC, when the user press request, but balance has not been fetched.
It will show balance is 0 please purchase minutes.
Fix:
Use user default to store balance when the user open the app and update user default when balance is updated

2.
Bug:
Ondisconnect function did not cancel when chatVC disappear
Fix:
Cancel disconnectOperation when chatVC disappear

3.
Bug: Both summaryVC and tutorConnectedVC  can push chatVC
Fix:
In tutorconnectedVC, set a flag call didTutorReady = false, in summaryVC, when observe status ==2,
set this flag to true, in viewWillAppear, if this flag is true, push chatVC, else, add observer to status