Update to FirebaseAuth 4.3.0 (was 4.0.0)
Because if the user disable notification when he first opens the app, he can't get the verification code.
Silent APNs notifications: When you sign in a user with their phone number for the first time on a device,
Firebase Authentication sends a token to the device using a silent push notification.
If your app successfully receives the notification from Firebase, phone number sign-in can proceed.
reCAPTCHA verification: In the event that sending or receiving a silent push notification is not possible,
 such as when the user has disabled background refresh for your app, or when testing your app on an iOS simulator,
 Firebase Authentication uses reCAPTCHA verification to complete the phone sign-in flow.
The reCAPTCHA challenge can often be completed without the user having to solve anything.